Fixed a previous mistake in the set methods. Changed variables to double.

+++ app/src/main/java/com/guzman/model/GeoLocationObjectModel.java
@@ -20,7 +20,7 @@ public class GeoLocationObjectModel implements Serializable
         return longitude;
     }
 
-    public void setLongitude(float longitude)
+    public void setLongitude(double longitude)
     {
         this.longitude = longitude;
     }
@@ -30,7 +30,7 @@ public class GeoLocationObjectModel implements Serializable
         return latitude;
     }
 
-    public void setLatitude(float latitude)
+    public void setLatitude(double latitude)
     {
         this.latitude = latitude;
     }
